# docs/swagger.yaml
basePath: /
definitions:
  PaymentGateway_internal_presentation_rest_dto.GetPaymentResponse:
    properties:
      amount:
        type: integer
      card_number_last_four:
        type: string
      currency:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
      id:
        type: string
      payment_status:
        type: string
    type: object
  PaymentGateway_internal_presentation_rest_dto.PostPaymentRequest:
    properties:
      amount:
        description: 'Validation: Required, must be greater than 0 (minor units/cents).'
        minimum: 1
        type: integer
      card_number:
        description: 'Validation: Required, must be numbers only, between 14 and 19
          digits.'
        type: string
      currency:
        description: 'Validation: Required, exactly 3 alphabetical characters (e.g.,
          "USD", "GBP").'
        type: string
      cvv:
        description: 'Validation: Required, 3 to 4 numeric digits.'
        maxLength: 4
        minLength: 3
        type: string
      expiry_month:
        description: 'Validation: Required, standard 1-12 months.'
        maximum: 12
        minimum: 1
        type: integer
      expiry_year:
        description: 'Validation: Required, must be at least the current year'
        minimum: 2026
        type: integer
    required:
    - amount
    - card_number
    - currency
    - cvv
    - expiry_month
    - expiry_year
    type: object
  PaymentGateway_internal_presentation_rest_dto.PostPaymentResponse:
    properties:
      amount:
        type: integer
      card_number_last_four:
        type: string
      currency:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
      id:
        type: string
      payment_status:
        description: '"Authorized", "Declined", "Rejected"'
        type: string
    type: object
host: localhost:8090
info:
  contact:
    name: API Support
  description: A highly available, PCI-compliant Payment Gateway API for processing
    and retrieving payments.
  title: Payment Gateway API
  version: "1.0"
paths:
  /v1/payments:
    post:
      consumes:
      - application/json
      description: Authorizes a payment through the acquiring bank simulator. Requires
        a unique Idempotency-Key header for safe retries.
      parameters:
      - description: Idempotency Key (UUID) for safe retries
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Payment Request details (Card info, amount, currency)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PaymentGateway_internal_presentation_rest_dto.PostPaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/PaymentGateway_internal_presentation_rest_dto.PostPaymentResponse'
        "400":
          description: Bad Request - Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict - Idempotency Key already used for a different request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Process a payment
      tags:
      - Payments
  /v1/payments/{id}:
    get:
      description: Gets the details, masked card numbers, and status of a previously
        processed payment using its unique ID.
      parameters:
      - description: Unique Payment ID (e.g., pay_123xyz)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PaymentGateway_internal_presentation_rest_dto.GetPaymentResponse'
        "404":
          description: Not Found - Payment does not exist
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve a payment
      tags:
      - Payments
swagger: "2.0"